requirements.txt
pandas>=1.5.0
numpy>=1.21.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
requests>=2.26.0
schedule>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.19.0 